third_party/WebKit/LayoutTests/external/wpt/web-animations/timing-model/animations/play-states-expected.txt

This is a testharness.js-based test.
PASS reports 'idle' for an animation with an unresolved current time and no pending tasks
FAIL reports 'paused' for an animation with a pending pause task assert_equals: expected "paused" but got "pending"
PASS reports 'paused' for an animation with a resolved current time and unresolved start time
PASS reports 'running' for an animation with a resolved start time and current time
PASS reports 'finished' when playback rate > 0 and current time = target effect end
PASS reports 'running' when playback rate = 0 and current time = target effect end
FAIL reports 'running' when playback rate < 0 and current time = target effect end assert_equals: expected "running" but got "pending"
PASS reports 'running' when playback rate > 0 and current time = 0
PASS reports 'running' when playback rate = 0 and current time = 0
FAIL reports 'finished' when playback rate < 0 and current time = 0 assert_equals: expected "finished" but got "pending"
FAIL reports 'finished' when playback rate > 0 and current time = target effect end and there is a pending play task assert_equals: expected "finished" but got "pending"
FAIL reports 'running' when playback rate > 0 and current time < target effect end and there is a pending play task assert_equals: expected "running" but got "pending"
FAIL reports 'running' for a play-pending animation assert_equals: expected "running" but got "pending"
FAIL reports 'paused' for a pause-pending animation assert_equals: expected "paused" but got "pending"
FAIL reports 'finished' for a finished-pending animation assert_equals: expected "finished" but got "pending"
FAIL reports the play state based on the pending playback rate animation.updatePlaybackRate is not a function
Harness: the test ran to completion.
